Final changes

+++ src/wikiSpeak/Main.java
@@ -19,6 +19,7 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.Scene;
 import javafx.scene.control.ListView;
+import javafx.scene.control.ProgressIndicator;
 import javafx.scene.control.TextInputDialog;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
@@ -26,6 +27,7 @@ import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.HBox;
+import javafx.scene.layout.Priority;
 import javafx.scene.layout.Region;
 import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
@@ -36,7 +38,9 @@ public class Main extends Application{
 	private Button _deleteButton;
 	private Button _playButton;
 	private ListView<String> _creationListView;
+	private ProgressIndicator _progress;
 	private BorderPane _mainPane;
+	private HBox _botPane;
 	private File _creationsFolder;
 	private ExecutorService _team = Executors.newSingleThreadExecutor();
 	
@@ -49,6 +53,9 @@ public class Main extends Application{
 		_createButton = new Button("Create");
 		_deleteButton = new Button("Delete");
 		_playButton = new Button("Play");
+		
+		_progress = new ProgressIndicator();
+		_progress.setMaxSize(26,26);
 		
 		_creationsFolder = new File("./Creations");
 		_creationsFolder.mkdirs();
@@ -75,11 +82,13 @@ public class Main extends Application{
 						String wordToSearch = searchDialog.getEditor().getText();
 						SearchTask searchJob = new SearchTask(wordToSearch);
 						_team.submit(searchJob);
+						_botPane.getChildren().add(_progress);
 						searchJob.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
 							
 							@Override
 							public void handle(WorkerStateEvent event) {
 								try {
+									_botPane.getChildren().remove(_progress);
 									List<String> rawSentences = searchJob.get();
 										if(rawSentences != null) {
 											CreationDialog creationDetailsDialog = new CreationDialog(AlertType.CONFIRMATION, rawSentences);
@@ -91,7 +100,6 @@ public class Main extends Application{
 												@Override
 												public void handle(ActionEvent e) {
 													String creationName = creationDetailsDialog.getTextField().getText();
-													System.out.println("creation name is - (" + creationName + ")");
 													//Check if entered name is invalid (must only be letters, digits, hyphen or underscore)
 													if((creationName.matches("^[a-zA-Z0-9_-]*$")) | !(creationName.trim().equals("")) | (creationName.equals(null))) {
 														
@@ -121,15 +129,7 @@ public class Main extends Application{
 																	if(result.get() == ButtonType.OK) {
 																		File file = new File(_creationsFolder.toString() + "/" + creationName + ".mp4");
 																		file.delete();
-																		_team.submit(createJob);
-																		createJob.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
-							
-																			@Override
-																			public void handle(WorkerStateEvent event) {
-																				refreshListView();
-																			}
-																			
-																		});
+																		runCreateCreationTask(createJob);
 																	}
 																}
 																
@@ -137,15 +137,7 @@ public class Main extends Application{
 															Platform.runLater(fileAlreadyExists);
 														}
 														else {
-															_team.submit(createJob);
-															createJob.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
-				
-																@Override
-																public void handle(WorkerStateEvent event) {
-																	refreshListView();
-																}
-																
-															});
+															runCreateCreationTask(createJob);
 														}
 													}	
 													else {
@@ -188,20 +180,21 @@ public class Main extends Application{
 			public void handle(ActionEvent arg0) {
 				String selectedItem = _creationListView.getSelectionModel().getSelectedItem();
 				
-				//Create a dialog for user to confirm their delete selection
-				Alert confirmAlert = new Alert(Alert.AlertType.CONFIRMATION);
-				((Button) confirmAlert.getDialogPane().lookupButton(ButtonType.OK)).setText("Yes");
-				confirmAlert.setTitle("Confirmation");
-				confirmAlert.setHeaderText(null);
-				confirmAlert.setContentText("You are about to delete " + selectedItem + ". Are you sure?");
-				Optional<ButtonType> result = confirmAlert.showAndWait();
-				if(result.get() == ButtonType.OK) {
-					File file = new File(_creationsFolder.toString() + "/" + selectedItem);
-					file.delete();
+				if(selectedItem != null) {
+					//Create a dialog for user to confirm their delete selection
+					Alert confirmAlert = new Alert(Alert.AlertType.CONFIRMATION);
+					((Button) confirmAlert.getDialogPane().lookupButton(ButtonType.OK)).setText("Yes");
+					confirmAlert.setTitle("Confirmation");
+					confirmAlert.setHeaderText(null);
+					confirmAlert.setContentText("You are about to delete " + selectedItem + ". Are you sure?");
+					Optional<ButtonType> result = confirmAlert.showAndWait();
+					if(result.get() == ButtonType.OK) {
+						File file = new File(_creationsFolder.toString() + "/" + selectedItem);
+						file.delete();
+					}
+					
+					refreshListView();
 				}
-				
-				refreshListView();
-				
 			}
 			
 		});
@@ -223,9 +216,12 @@ public class Main extends Application{
 			
 		});
 		
+		//Make the progress indicator anchor to right side
+		Region region = new Region();
+		HBox.setHgrow(region, Priority.ALWAYS);
 		
-		HBox botPane = new HBox(10, _createButton, _deleteButton, _playButton);
-		_mainPane.setBottom(botPane);
+		_botPane = new HBox(10, _createButton, _deleteButton, _playButton, region);
+		_mainPane.setBottom(_botPane);
 		
 		Scene scene = new Scene(_mainPane, 500,300);
 		primaryStage.setScene(scene);
@@ -237,6 +233,28 @@ public class Main extends Application{
 			public void handle(WindowEvent e) {
 				Platform.exit();
 				System.exit(0);
+			}
+			
+		});
+	}
+	
+	private void runCreateCreationTask(MakeCreationTask createJob) {
+		_team.submit(createJob);
+		_botPane.getChildren().add(_progress);
+		createJob.setOnSucceeded(new EventHandler<WorkerStateEvent>() {
+
+			@Override
+			public void handle(WorkerStateEvent event) {
+				Runnable r = new Runnable() {
+
+					@Override
+					public void run() {
+						_botPane.getChildren().remove(_progress);
+					}
+					
+				};
+				Platform.runLater(r);
+				refreshListView();
 			}
 			
 		});
